Added funcionalities of creating and editing the entity Classes
Currently there is a problem related to the ListView that displays classes when adding a new one, it does not update. However if you check CadastroActivity.java, which has a Spinner which gets all the Classes from the Classes entity you can see that the new classes were added successfully.